Initial commit for fixing K20 UART
Fixes a race condition that causes read buffer corruption
Use K20 defines for register manipulation
Increase K20 buffer sizes

+++ interface/Common/src/main.c
@@ -164,7 +164,7 @@ void main_disable_debug_event(void) {
 
 __task void serial_process() {
     uint8_t data[SIZE_DATA];
-    uint8_t len_data = 0;
+    int32_t len_data = 0;
 
     while (1) {
